// src/app/modules/admin/driver/interfaces/driver.interfaces.ts
export interface Driver {
    id?:               number;
    name:              string;
    last_name:         string;
    birth_date:        string;
    curp:              string;
    address:           string;
    monthly_salary:    number;
    driving_license:   string;
    registration_date: string;
}